refactor(ui): split render module into focused submodules
